Split the DAO into per-domain classes and defer polling until schema ready
Break the single dao/DAO.ts class into one DAO per table-group, each in its own
file: UserDAO, ChatDAO, GithubDAO, GameUpdateDAO, ScheduleDAO, RsvpDAO and
RollcallPlayerDAO. The shared telegram_user upsert moves to a standalone
ensureTelegramUser() helper (dao/telegramUser.ts); domain types move to their
owning DAO file. Callers instantiate only the DAOs they use. Behaviour is
unchanged (method bodies moved verbatim).

Also fixes a startup race flagged in review (PR #28): the Telegram bot now
constructs with polling disabled and only calls bot.startPolling() inside the
ensureSchema().then(), so no message/callback handler can hit a DAO method
before the MariaDB tables exist.

Verified: tsc + build clean; acceptance suite (14 features / 38 scenarios /
179 steps) and the migration + steam-backend integration test all pass.
